add overloaded method for calcFeetAndInchesToCentimeters with one parameter

+++ src/methodOverloadingDemo/CalculateCentimeters.java
@@ -40,7 +40,7 @@ public class CalculateCentimeters {
     // overloading method calcFeetAndInchesToCentimeters()
     public static double calcFeetAndInchesToCentimeters(double inches2) {
         if (inches2 >= 0) {
-            double feet2 = inches2 / 12;
+            double feet2 = (int) inches2 / 12;
             double remainingInches = inches2 % 12;
             // calling main method in overloaded one
             return calcFeetAndInchesToCentimeters(feet2, remainingInches);
